API Shifted to new directory API
