+++ config/api_router.py
@@ -1,12 +1,9 @@
 from django.conf import settings
 from rest_framework.routers import DefaultRouter, SimpleRouter
 
-from fight_covid19.maps.api.views import (
-    HealthEntryViewSet,
-    CoronaVirusCasesViewSet,
-    HealthStatisticsViewSet,
-)
-from fight_covid19.users.api.views import UserViewSet
+from fight_covid19.api.coronacases.views import CoronaVirusCasesViewSet
+from fight_covid19.api.hoi.views import HealthEntryViewSet, HealthStatisticsViewSet
+from fight_covid19.api.users.views import UserViewSet
 
 if settings.DEBUG:
     router = DefaultRouter()
