+++ fight_covid19/api/hoi/serializers.py
@@ -4,6 +4,8 @@ from fight_covid19.maps.models import HealthEntry
 
 
 class HealthEntryFormSerializer(serializers.HyperlinkedModelSerializer):
+    """Serializers for Health Entry"""
+
     class Meta:
         model = HealthEntry
         fields = [
@@ -16,6 +18,7 @@ class HealthEntryFormSerializer(serializers.HyperlinkedModelSerializer):
             "self_quarantine",
             "latitude",
             "longitude",
+            "unique_id",
         ]
 
 
@@ -35,5 +38,6 @@ class HealthEntrySerializer(serializers.HyperlinkedModelSerializer):
             "self_quarantine",
             "latitude",
             "longitude",
+            "unique_id",
             "creation_timestamp",
         ]
